Refactor: Add wati.io API integration for sending template messages
The code changes in this commit add integration with the wati.io API to send template messages. It includes the addition of the `sendTemplateMessagesAPI` constant to store the API endpoint URL. The `expandReceivers` function is also added to handle expanding receivers with multiple WhatsApp numbers. This commit enables the functionality to send template messages using the wati.io API.